Java Advanced - Exceptions-  02-Exception-Handling.md
Lubo please look into the Problems and Solutions here. We don't have solution videos and the tasks are not actual tasks, just text.

I've rewritten the last problem, it was not correct at all but again, we don't have a solution for it.  Should I just write the solutions? How are the students going to check if the tasks are correct?